requirements.txt
pandas==2.0.2
python-dotenv==1.0.0
requests==2.22.0
schedule==1.2.0
tabulate==0.9.0
pyyaml==6.0